+++ .idea/sqldialects.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="SqlDialectMappings">
-    <file url="file://$PROJECT_DIR$/src/main/java/nl/han/oose/lindt/maarten/datasource/DAO/PlaylistDAO.java" dialect="GenericSQL" />
+    <file url="file://$PROJECT_DIR$/src/main/java/nl/han/oose/lindt/maarten/datasource/dao/PlaylistDAO.java" dialect="GenericSQL" />
   </component>
 </project>
