+++ src/main/java/nl/han/oose/lindt/maarten/datasource/dao/FailedConnectionException.java
@@ -1,4 +1,4 @@
-package nl.han.oose.lindt.maarten.datasource.DAO;
+package nl.han.oose.lindt.maarten.datasource.dao;
 
 public class FailedConnectionException extends RuntimeException {
 }
