+++ src/main/java/nl/han/oose/lindt/maarten/datasource/dao/FailedResultsetReadingException.java
@@ -1,4 +1,4 @@
-package nl.han.oose.lindt.maarten.datasource.DAO;
+package nl.han.oose.lindt.maarten.datasource.dao;
 
 public class FailedResultsetReadingException extends RuntimeException {
 }
