+++ src/main/java/nl/han/oose/lindt/maarten/resources/exceptionmappers/FailedConnectionExceptionMapper.java
@@ -1,6 +1,6 @@
-package nl.han.oose.lindt.maarten.resources.exceptionMappers;
+package nl.han.oose.lindt.maarten.resources.exceptionmappers;
 
-import nl.han.oose.lindt.maarten.datasource.DAO.FailedConnectionException;
+import nl.han.oose.lindt.maarten.datasource.dao.FailedConnectionException;
 
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
