+++ src/main/java/nl/han/oose/lindt/maarten/services/HardcodedTrackService.java
@@ -61,6 +61,6 @@ public class HardcodedTrackService implements TrackService {
                 }
             }
         }
-        return checkOKE;
+
     }
 }
